Cambios Finales Funcionales

+++ slnSistemaCitas/CapaNegocio/clsN_Promociones.cs
@@ -13,7 +13,7 @@ namespace CapaNegocio
     {
         public int idP { get; set; }
         public string nombreP { get; set; }
-        public decimal descuentoP { get; set; }
+        public int descuentoP { get; set; }
 
         clsD_Promociones D_Promociones = new clsD_Promociones();
 
@@ -27,14 +27,14 @@ namespace CapaNegocio
             return D_Promociones.consultaPromociones();
         }
 
-        public bool agregarPromocion(string nombre, decimal descuento)
+        public bool agregarPromocion(string nombre, int descuento)
         {
             nombreP = nombre;
             descuentoP = descuento;
             return (D_Promociones.agregarPromocion(nombreP, descuentoP));
         }
 
-        public bool modificarPromocion(int id, string nombre, decimal descuento)
+        public bool modificarPromocion(int id, string nombre, int descuento)
         {
             idP = id;
             nombreP = nombre;
